reorganize file structure a little
