+++ frontend/src/features/qa/page-list/helpers/groupBy.ts
@@ -1,4 +1,4 @@
-import type { OrderBy } from "..";
+import { type OrderBy } from "../page-list";
 
 import { severityFromMatch } from "./severity";
